[Professor-review test] já dá notas às perguntas mas ainda não está terminado

+++ etoile-play/app/models/manytomany/UserTest.java
@@ -23,7 +23,7 @@ public class UserTest extends Model{
 	public Long id;
 
 	@Constraints.Required
-	public int reputationAsAstudent;
+	public double reputationAsAstudent;
 	
 	@Constraints.Required
 	public int reputationAsAMarker;
